refactor: rename context by provider
